+++ src/logger/builder.rs
@@ -22,7 +22,7 @@ use crate::Append;
 use crate::Diagnostic;
 use crate::Filter;
 
-/// Creates a new empty [`Builder`] instance for configuring log dispatching.
+/// Creates a new empty [`LoggerBuilder`] instance for configuring log dispatching.
 ///
 /// # Examples
 ///
@@ -33,11 +33,11 @@ use crate::Filter;
 ///     .dispatch(|d| d.append(append::Stderr::default()))
 ///     .apply();
 /// ```
-pub fn builder() -> Builder {
-    Builder::new()
+pub fn builder() -> LoggerBuilder {
+    LoggerBuilder::new()
 }
 
-/// Creates a [`Builder`] with a default [`append::Stdout`] appender and an [`env_filter`](https://crates.io/crates/env_filter)
+/// Creates a [`LoggerBuilder`] with a default [`append::Stdout`] appender and an [`env_filter`](https://crates.io/crates/env_filter)
 /// respecting `RUST_LOG`.
 ///
 /// # Examples
@@ -46,14 +46,14 @@ pub fn builder() -> Builder {
 /// logforth::stdout().apply();
 /// log::error!("This error will be logged to stdout.");
 /// ```
-pub fn stdout() -> Builder {
+pub fn stdout() -> LoggerBuilder {
     builder().dispatch(|d| {
         d.filter(EnvFilter::from_default_env())
             .append(append::Stdout::default())
     })
 }
 
-/// Creates a [`Builder`] with a default [`append::Stderr`] appender and an [`env_filter`](https://crates.io/crates/env_filter)
+/// Creates a [`LoggerBuilder`] with a default [`append::Stderr`] appender and an [`env_filter`](https://crates.io/crates/env_filter)
 /// respecting `RUST_LOG`.
 ///
 /// # Examples
@@ -62,7 +62,7 @@ pub fn stdout() -> Builder {
 /// logforth::stderr().apply();
 /// log::info!("This info will be logged to stderr.");
 /// ```
-pub fn stderr() -> Builder {
+pub fn stderr() -> LoggerBuilder {
     builder().dispatch(|d| {
         d.filter(EnvFilter::from_default_env())
             .append(append::Stderr::default())
@@ -80,25 +80,19 @@ pub fn stderr() -> Builder {
 ///     .dispatch(|d| d.append(append::Stdout::default()))
 ///     .apply();
 /// ```
-#[must_use = "call `apply` to set the global logger"]
+#[must_use = "call `apply` to set the global logger or `build` to construct a logger instance"]
 #[derive(Debug)]
-pub struct Builder {
+pub struct LoggerBuilder {
     // stashed dispatches
     dispatches: Vec<Dispatch>,
-
-    // default to trace - we need this because the global default is OFF
-    max_level: LevelFilter,
 }
 
-impl Builder {
+impl LoggerBuilder {
     fn new() -> Self {
-        Builder {
-            dispatches: vec![],
-            max_level: LevelFilter::Trace,
-        }
+        LoggerBuilder { dispatches: vec![] }
     }
 
-    /// Registers a new dispatch with the [`Builder`].
+    /// Registers a new dispatch with the [`LoggerBuilder`].
     ///
     /// # Examples
     ///
@@ -117,26 +111,27 @@ impl Builder {
         self
     }
 
-    /// Sets the global maximum log level. Default to [`LevelFilter::Trace`].
-    ///
-    /// This will be passed to `log::set_max_level()`.
+    /// Build the [`Logger`].
     ///
     /// # Examples
     ///
     /// ```
-    /// logforth::builder()
-    ///     .max_level(log::LevelFilter::Warn)
-    ///     .apply();
+    /// let l = logforth::builder().build();
+    /// log::error!(logger: l, "Hello error!");
     /// ```
-    pub fn max_level(mut self, max_level: LevelFilter) -> Self {
-        self.max_level = max_level;
-        self
+    pub fn build(self) -> Logger {
+        Logger::new(self.dispatches)
     }
 
     /// Sets up the global logger with all the configured dispatches.
     ///
     /// This should be called early in the execution of a Rust program. Any log events that occur
     /// before initialization will be ignored.
+    ///
+    /// This will set the global maximum log level to [`LevelFilter::Trace`]. To override this,
+    /// call [`log::set_max_level`] after this function. Alternatively, you can obtain a [`Logger`]
+    /// instance by calling [`LoggerBuilder::build`], and then call [`log::set_boxed_logger`]
+    /// manually.
     ///
     /// # Errors
     ///
@@ -151,9 +146,9 @@ impl Builder {
     /// }
     /// ```
     pub fn try_apply(self) -> Result<(), log::SetLoggerError> {
-        let logger = Logger::new(self.dispatches);
+        let logger = self.build();
         log::set_boxed_logger(Box::new(logger))?;
-        log::set_max_level(self.max_level);
+        log::set_max_level(LevelFilter::Trace);
         Ok(())
     }
 
@@ -161,6 +156,11 @@ impl Builder {
     ///
     /// This function will panic if it is called more than once, or if another library has already
     /// initialized a global logger.
+    ///
+    /// This function will set the global maximum log level to [`LevelFilter::Trace`]. To override
+    /// this, call [`log::set_max_level`] after this function. Alternatively, you can obtain a
+    /// [`Logger`] instance by calling [`LoggerBuilder::build`], and then call
+    /// [`log::set_boxed_logger`] manually.
     ///
     /// # Panics
     ///